move all extension functions to data module
