+++ app/src/main/java/com/my/kizzy/MainActivity.kt
@@ -47,7 +47,7 @@ import com.my.kizzy.ui.theme.KizzyTheme
 import com.my.kizzy.ui.theme.LocalDarkTheme
 import com.my.kizzy.ui.theme.LocalDynamicColorSwitch
 import com.my.kizzy.ui.theme.SettingsProvider
-import com.my.kizzy.utils.hasUsageAccess
+import com.my.kizzy.data.utils.hasUsageAccess
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.runBlocking
